Add files via upload
This query resolves the issue in sorting string values in SQL, usually SQL takes values in order like 1, 10, 11, 12, ... , 2, 20, 21 and so on. I solved this and tested on my university data of different batches i.e., 18TL, 19TL, 19BM, 20AR. If the roll no are containing values 18-17tl80 it will sort them as well.